+++ vector/src/main/java/im/vector/riotx/core/di/FragmentModule.kt
@@ -78,6 +78,7 @@ import im.vector.riotx.features.roommemberprofile.RoomMemberProfileFragment
 import im.vector.riotx.features.roommemberprofile.devices.DeviceListFragment
 import im.vector.riotx.features.roommemberprofile.devices.DeviceTrustInfoActionFragment
 import im.vector.riotx.features.roomprofile.RoomProfileFragment
+import im.vector.riotx.features.roomprofile.banned.RoomBannedMemberListFragment
 import im.vector.riotx.features.roomprofile.members.RoomMemberListFragment
 import im.vector.riotx.features.roomprofile.settings.RoomSettingsFragment
 import im.vector.riotx.features.roomprofile.uploads.RoomUploadsFragment
@@ -540,4 +541,9 @@ interface FragmentModule {
     @IntoMap
     @FragmentKey(PinFragment::class)
     fun bindPinFragment(fragment: PinFragment): Fragment
+
+    @Binds
+    @IntoMap
+    @FragmentKey(RoomBannedMemberListFragment::class)
+    fun bindRoomBannedMemberListFragment(fragment: RoomBannedMemberListFragment): Fragment
 }