+++ vector/src/main/java/im/vector/riotx/features/signout/soft/SoftLogoutController.kt
@@ -28,7 +28,6 @@ import im.vector.riotx.core.resources.StringProvider
 import im.vector.riotx.features.login.LoginMode
 import im.vector.riotx.features.signout.soft.epoxy.loginCenterButtonItem
 import im.vector.riotx.features.signout.soft.epoxy.loginErrorWithRetryItem
-import im.vector.riotx.features.signout.soft.epoxy.loginHeaderItem
 import im.vector.riotx.features.signout.soft.epoxy.loginPasswordFormItem
 import im.vector.riotx.features.signout.soft.epoxy.loginRedButtonItem
 import im.vector.riotx.features.signout.soft.epoxy.loginTextItem
@@ -65,9 +64,6 @@ class SoftLogoutController @Inject constructor(
     }
 
     private fun buildHeader(state: SoftLogoutViewState) {
-        loginHeaderItem {
-            id("header")
-        }
         loginTitleItem {
             id("title")
             text(stringProvider.getString(R.string.soft_logout_title))